Improve /lagstat command (#43)
The lagstat command `/lagstat` has been rewritten entirely. The command now shows how many "tiny" lag spikes (between 10mx and 70ms) that aren't noticeable individually, but could contribute to the game running at a lower FPS than it should, and "big" lag spikes that are more visible (>70ms).

Basically the server times how long it takes from sending a response, to getting back the next set of inputs for that user. Note that even testing locally I was getting about ~18ms when ideally the target is 16.66ms.

In the future it should be possible to detect when users across the server are collectively lagging (most likely meaning a network issue close to the server) and provide both network health metrics for users and the server. The server could nudge users to use a specific fake ping.

This will all take more experimentation but at least this is more useful than the current command. The new version is enabled by default and can be disabled by setting `server.improvedLagstatEnabled = false` in `conf/emulinker.config`.